Rework gui_list_view_ex.
Working with CStr is not practical, and expecting the user to have to manipulate them is
even less. Users are very likely to need to make a wrapper in order use this interface,
so rework it to make it better to use.
Expect a iterator of String for the text, this is more flexible than using a array and
doesn't complicate the implementation as we will have to convert it to a list of CStr anyway.